refactor: move builtins to authly-domain
