+++ lib/authly-domain/src/lib.rs
@@ -1,2 +1,4 @@
+pub mod builtins;
 pub mod ctx;
+pub mod directory;
 pub mod id;
