+++ src/db/init_db.rs
@@ -4,37 +4,17 @@ use std::{borrow::Cow, collections::HashMap, time::Duration};
 
 use authly_common::id::{AttrId, DirectoryId, Id128DynamicArrayConv, PropId, ServiceId};
 use authly_db::{param::ToBlob, params, Db, DbError, DbResult, FromRow};
-use authly_domain::id::{BuiltinAttr, BuiltinProp};
+use authly_domain::{
+    builtins::Builtins,
+    directory::DirKey,
+    id::{BuiltinAttr, BuiltinProp},
+};
 use fnv::FnvHashMap;
 use hiqlite::StmtIndex;
 use indoc::indoc;
 use tracing::info;
 
-use crate::{directory::DirKey, IsLeaderDb};
-
-pub struct Builtins {
-    pub authly_dir_key: DirKey,
-    pub authly_namespace_key: i64,
-
-    prop_keys: FnvHashMap<BuiltinProp, i64>,
-    attr_keys: FnvHashMap<BuiltinAttr, i64>,
-}
-
-impl Builtins {
-    pub fn prop_key(&self, prop: BuiltinProp) -> i64 {
-        self.prop_keys
-            .get(&prop)
-            .copied()
-            .expect("builtin prop without key")
-    }
-
-    pub fn attr_key(&self, attr: BuiltinAttr) -> i64 {
-        self.attr_keys
-            .get(&attr)
-            .copied()
-            .expect("builtin attr without key")
-    }
-}
+use crate::IsLeaderDb;
 
 #[derive(Debug)]
 enum InitDbError {
