+++ src/persona_directory.rs
@@ -1,6 +1,6 @@
 use authly_common::id::PersonaId;
 use authly_db::{DbError, DidInsert};
-use authly_domain::{ctx::GetDb, id::BuiltinProp};
+use authly_domain::{ctx::GetDb, directory::DirKey, id::BuiltinProp};
 use tracing::info;
 
 use crate::{
@@ -10,7 +10,6 @@ use crate::{
         entity_db::{self, OverwritePersonaId},
         object_db,
     },
-    directory::DirKey,
 };
 
 // A persona whose source is a 3rd-party directory (OAuth/LDAP/etc)
